requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
asgiref==3.4.1
async-exit-stack==1.0.1
async-generator==1.10
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
distlib==0.3.2
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.66.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
numpy==1.21.0
orjson==3.6.0
pandas==1.3.0
promise==2.3
psycopg2==2.9.1
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
Rx==1.6.1
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
Werkzeug==2.0.1
